Unit test refactor & made OutcomeTestHelper for both outcome test files
